+++ src/main/java/com/ita/if103java/ims/service/TransactionService.java
@@ -2,8 +2,6 @@ package com.ita.if103java.ims.service;
 
 import com.ita.if103java.ims.dto.TransactionDto;
 
-import java.math.BigInteger;
-
 public interface TransactionService {
-    TransactionDto findById(BigInteger id);
+    TransactionDto findById(Long id);
 }
